Wed, Nov 8, 2023, 8:32 PM -08:00

+++ 2324/chapter2/examples/calc/calc.c
@@ -18,8 +18,8 @@ int main(int argc, string argv[])
         return 1;
     }
 
-    float num1 = atof(argv[1]);
-    float num2 = atof(argv[3]);
+    double num1 = atof(argv[1]);
+    double num2 = atof(argv[3]);
     char operator= argv[2][0]; // assign the first character of the third  command line argument (index: 2) as the operator -->
                                // needed as you cant firectly assign string to char
 
@@ -75,7 +75,7 @@ int main(int argc, string argv[])
     }
 }
 
-float mod(float num1, float num2)
+double mod(float num1, float num2)
 {
     if (num2 == 0)
     {
